src/orq_ai_sdk/openai_agents_instrumentation/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
googleapis-common-protos==1.70.0
griffe==1.14.0
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.17.0
openai==1.109.1
openai-agents==0.3.3
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
protobuf==6.32.1
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
wrapt==1.17.3
zipp==3.23.0
